% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{arrow_lighting}
\alias{arrow_lighting}
\title{Highlight arrows for steady state events}
\usage{
arrow_lighting(edges, simulation, counter, session)
}
\arguments{
\item{edges}{A dataframe of edges provided by \link{generate_edges}.}

\item{simulation}{Which disease is currently selected. See \link{extract_running_sim}.}

\item{counter}{To determine which notification to display. We expect a counter
returned by the \link{networkCaPO4} module.}

\item{session}{Session object.}
}
\description{
Use inside in the \link{networkCaPO4}. Nothing is returned
except that the network is updated via \link[visNetwork]{visNetworkProxy}.
}
