% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseaseModule.R
\name{casestudiesCheckBox}
\alias{casestudiesCheckBox}
\title{Create a checkbox for \link{casestudiesSelectUi}}
\usage{
casestudiesCheckBox(inputId, label)
}
\arguments{
\item{inputId}{Checkbox Input id.}

\item{label}{Checkbox label.}
}
\description{
Create a \link[shinyWidgets]{prettyCheckbox}.
}
