% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseaseModule.R
\name{casestudiesSelect}
\alias{casestudiesSelect}
\title{Create a casestudies selector server logic}
\usage{
casestudiesSelect(input, output, session)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny Outputs}

\item{session}{Session object.}
}
\description{
Only returns inputs associated with php1, hypopara, hypoD3
}
