% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseaseModule.R
\name{casestudiesSelectUi}
\alias{casestudiesSelectUi}
\title{Create a casestudies selector UI module}
\usage{
casestudiesSelectUi(id)
}
\arguments{
\item{id}{module id.}
}
\description{
Contains php1, hypopara, hypoD3
}
