% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseaseModule.R
\name{diseaseCheckBox}
\alias{diseaseCheckBox}
\title{Create a checkbox for \link{diseaseSelectUi}}
\usage{
diseaseCheckBox(inputId, label)
}
\arguments{
\item{inputId}{Checkbox Input id.}

\item{label}{Checkbox label.}
}
\description{
Create a \link[shinyWidgets]{prettyCheckbox}.
}
