% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{extract_running_sim}
\alias{extract_running_sim}
\title{Extract the current running simulation}
\usage{
extract_running_sim(diseases)
}
\arguments{
\item{diseases}{Shiny input disease selector. See \link{diseaseSelect}.}
}
\description{
Simulations are currently php1, hypoD3 and hypopara. Takes
diseases as input given by the \link{diseaseSelect} module.
}
