% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullScreenModule.R
\name{fullScreen}
\alias{fullScreen}
\title{Create a fullScreen server logic}
\usage{
fullScreen(input, output, session)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny Outputs}

\item{session}{Session object.}
}
\description{
Nothing is contained inside for now...
}
