% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullScreenModule.R
\name{fullScreenUI}
\alias{fullScreenUI}
\title{Create a fullScreen UI module}
\usage{
fullScreenUI(id)
}
\arguments{
\item{id}{module id.}
}
\description{
Trigger a fullScreen mode. Based on
https://stackoverflow.com/questions/42371164/how-to-run-r-shiny-app-in-full-sized-window
}
