% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{generate_edges}
\alias{generate_edges}
\title{CaPO4 Edges Generator}
\usage{
generate_edges(components, organs, regulations, diseases,
  organs_nodes_size, hormones_nodes_size)
}
\arguments{
\item{components}{Shiny input CaPO4 component selector. See \link{networkOptions}.}

\item{organs}{Shiny input to toggle organs display. See \link{networkOptions}.}

\item{regulations}{Shiny input to toggle hormone display. See \link{networkOptions}.}

\item{diseases}{Shiny input disease selector. See \link{diseaseSelect}.}

\item{organs_nodes_size}{Shiny input for organs node size. See \link{networkOptions}.}

\item{hormones_nodes_size}{Shiny input for hormones node size. See \link{networkOptions}.}

\item{background}{Shiny input background selector. See \link{networkOptions}.}

\item{organs_edges_size}{Shiny input for organs edges size. See \link{networkOptions}.}

\item{hormones_edges_size}{Shiny input for hormones edges size. See \link{networkOptions}.}
}
\description{
Generate edges for the CaPO4 network
}
