% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{generate_network}
\alias{generate_network}
\title{CardioRenal Network Generator}
\usage{
generate_network(nodes, edges, usephysics = FALSE, isMobile)
}
\arguments{
\item{nodes}{A dataframe of nodes provided by \link{generate_nodes}.}

\item{edges}{A dataframe of edges provided by \link{generate_edges}.}

\item{usephysics}{Whether to use physic. FALSE by default. A visNetwork API parameter.}

\item{isMobile}{Shiny input checking if the app is running on a cellphone/tablet.}
}
\description{
Create a CardioRenal network taking nodes and edges as inputs
}
