% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{generate_nodes}
\alias{generate_nodes}
\title{CaPO4 Nodes Generator}
\usage{
generate_nodes(
  components,
  organs,
  regulations,
  background,
  diseases,
  organs_nodes_size,
  hormones_nodes_size
)
}
\arguments{
\item{components}{Shiny input CaPO4 component selector. See \link{networkOptions}.}

\item{organs}{Shiny input to toggle organs display. See \link{networkOptions}.}

\item{regulations}{Shiny input to toggle hormone display. See \link{networkOptions}.}

\item{background}{Shiny input background selector. See \link{networkOptions}.}

\item{diseases}{Shiny input disease selector. See \link{diseaseSelect}.}

\item{organs_nodes_size}{Shiny input for organs node size. See \link{networkOptions}.}

\item{hormones_nodes_size}{Shiny input for hormones node size. See \link{networkOptions}..}
}
\description{
Generate nodes for the CaPO4 network
}
