% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{generate_notification}
\alias{generate_notification}
\title{Notifications Generator for CaPO4 animations}
\usage{
generate_notification(simulation, counter, allowed)
}
\arguments{
\item{simulation}{Which disease is currently selected. See \link{extract_running_sim}.}

\item{counter}{To determine which notification to display. We expect a counter
returned by the \link{networkCaPO4} module.}

\item{allowed}{Whether to allow simulations. Expect logical value. See \link{infos} module.}
}
\description{
Generate sequential notification as a function of the
selected diseases. All notifications are in the notifications.R file in the
inst/entry_level app folder. Used in the \link{infos} module.
}
