% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{generate_userFields}
\alias{generate_userFields}
\title{Generate user fields}
\usage{
generate_userFields(diseases, sliderDisease)
}
\arguments{
\item{diseases}{Shiny input disease selector. See \link{diseaseSelect}.}

\item{sliderDisease}{Shiny slider input related to the current disease severity.
See \link{plotBox}.}
}
\description{
Use inside in the \link{userInfo}. Function that helps in generating 4 users fields,
image, stat1, stat2 and stat3, so as to reinject them in the header userMenu
}
