% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossaryModule.R
\name{glossaryCaPO4}
\alias{glossaryCaPO4}
\title{CaPO4 glossary server module}
\usage{
glossaryCaPO4(input, output, session)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny Outputs}

\item{session}{Session object.}
}
\description{
Create a CaPO4 glossary
}
