% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpModule.R
\name{helpCaPO4}
\alias{helpCaPO4}
\title{Help server module}
\usage{
helpCaPO4(input, output, session)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny Outputs}

\item{session}{Session object.}
}
\description{
Create the help section
}
