% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoModule.R
\name{info}
\alias{info}
\title{Info server module}
\usage{
info(input, output, session)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny Outputs}

\item{session}{Session object.}
}
\description{
Create modals, alerts, ...
}
