% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoModule.R
\name{infos}
\alias{infos}
\title{Info server module}
\usage{
infos(input, output, session, diseases, animation_counter, regulations)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny Outputs}

\item{session}{Session object.}

\item{diseases}{Shiny input disease selector. See \link{diseaseSelect}.}

\item{animation_counter}{Give the current temporal state of the animation. See \link{networkCaPO4}.}

\item{regulations}{Shiny input to toggle hormone display. See \link{networkOptions}.}
}
\description{
Create modals, alerts, ...
}
