% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoModule.R
\name{infoSwitch}
\alias{infoSwitch}
\title{Create a switch input for \link{infosUi}}
\usage{
infoSwitch(inputId, label)
}
\arguments{
\item{inputId}{Checkbox Input id.}

\item{label}{Checkbox label.}
}
\description{
Create a \link[shinyWidgets]{prettySwitch}.
}
