% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{make_plot_hypopara}
\alias{make_plot_hypopara}
\title{Produce plots related to hypoparathyroidism (hypopara)}
\usage{
make_plot_hypopara(sliderVal)
}
\arguments{
\item{sliderVal}{Shiny slider input related to the current disease severity.
See \link{plotBox}.}
}
\description{
Use inside the \link{plotBox} module.
}
