% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videoModule.R
\name{myCarousel}
\alias{myCarousel}
\title{carousel container}
\usage{
myCarousel(
  ...,
  id,
  indicators = TRUE,
  width = 6,
  .list = NULL,
  data.interval = 5000,
  data.ride = "carousel"
)
}
\arguments{
\item{...}{Slot for \link{carouselItem}}

\item{id}{Carousel id. Must be unique.}

\item{indicators}{Whether to display left and right indicators.}

\item{width}{Carousel width. 6 by default.}

\item{.list}{Should you need to pass \link{carouselItem} via \link{lapply} or similar,
put these item here instead of passing them in ...}

\item{data.interval}{specify data-interval in ms. 5000ms by default, set to "false" to prevent automated animation of the slides.}

\item{data.ride}{specify data-ride. "carousel" by default.}
}
\description{
Creates a carousel. Adapted from shinydashboardplus to also allow control of the carousel animation
}
