% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkModule.R
\name{networkCaPO4}
\alias{networkCaPO4}
\title{CaPO4 Network server module}
\usage{
networkCaPO4(
  input,
  output,
  session,
  isMobile,
  components,
  organs,
  regulations,
  background,
  diseases,
  organs_nodes_size,
  hormones_nodes_size,
  organs_edges_size,
  hormones_edges_size,
  help
)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny Outputs}

\item{session}{Session object.}

\item{isMobile}{Shiny input checking if the app is running on a cellphone/tablet.}

\item{components}{Shiny input CaPO4 component selector. See \link{networkOptions}.}

\item{organs}{Shiny input to toggle organs display. See \link{networkOptions}.}

\item{regulations}{Shiny input to toggle hormone display. See \link{networkOptions}.}

\item{background}{Shiny input background selector. See \link{networkOptions}.}

\item{diseases}{Shiny input disease selector. See \link{diseaseSelect}.}

\item{organs_nodes_size}{Shiny input for organs node size. See \link{networkOptions}.}

\item{hormones_nodes_size}{Shiny input for hormones node size. See \link{networkOptions}.}

\item{organs_edges_size}{Shiny input for organs edges size. See \link{networkOptions}.}

\item{hormones_edges_size}{Shiny input for hormones edges size. See \link{networkOptions}.}

\item{help}{Help input.}
}
\description{
Create a CaPO4 network
}
