% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkOptionsModule.R
\name{networkOptions}
\alias{networkOptions}
\title{CaPO4 Network Options server module}
\usage{
networkOptions(input, output, session, mobile)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny Outputs}

\item{session}{Session object.}

\item{mobile}{Whether we are on cellphone/tablets or not. Slot for input$ismobile().}
}
\description{
Create a CaPO4 network options
}
