% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_CardioRenalSim.R
\name{run_CardioRenalSim}
\alias{run_CardioRenalSim}
\title{Launch the virtual patient simulator}
\usage{
run_CardioRenalSim(context = c("introduction", "virtual-patient"))
}
\arguments{
\item{context}{Choose between \code{c("introduction", "virtual-patient")}.}
}
\description{
Unleash the virtual patient simulator
}
\examples{
if (interactive()) {
 run_CardioRenalSim(context = "introduction")
 run_CardioRenalSim(context = "virtual-patient")
}
}
