% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skinSelectModule.R
\name{skinSelect}
\alias{skinSelect}
\title{Dashboard skin selector, server side}
\usage{
skinSelect(input, output, session)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny Outputs}

\item{session}{Session object.}
}
\description{
Select the shinydashboard skin you want
}
