% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videoModule.R
\name{video}
\alias{video}
\title{Create a video server logic}
\usage{
video(input, output, session)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny Outputs}

\item{session}{Session object.}
}
\description{
Nothing is contained inside for now...
}
