% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaPO4Sim-package.R
\docType{package}
\name{CaPO4Sim}
\alias{CaPO4Sim-package}
\alias{CaPO4Sim}
\title{CaPO4Sim}
\description{
Explore calcium (Ca) and phosphate (Pi) homeostasis with two novel 'Shiny' apps,
building upon on a previously published mathematical model written in C,
to ensure efficient computations. The underlying model is accessible
here <https://www.ncbi.nlm.nih.gov/pubmed/28747359>.
The first application explores the fundamentals of Ca-Pi homeostasis,
while the second provides interactive case studies for in-depth exploration of the topic,
thereby seeking to foster student engagement and an integrative understanding of Ca-Pi regulation.
These applications are hosted at <https://rinterface.com/AppsPhysiol.html>.
}
\author{
\strong{Maintainer}: David Granjon \email{dgranjon@ymail.com} [copyright holder]

Other contributors:
\itemize{
  \item Diane de Zélicourt [copyright holder]
  \item Vartan Kurtcuoglu [copyright holder]
  \item Olivier Bonny [copyright holder]
  \item François Verrey [copyright holder]
  \item  University of Lausanne [funder]
  \item  University of Zurich [funder]
  \item  Kidney NCCR.CH [funder]
  \item  The Interface Group (Hosting Group) [copyright holder]
  \item  RinteRface (R/HTML Templates) [copyright holder]
}

}
