% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{extract_running_sim}
\alias{extract_running_sim}
\title{Extract the current running simulation}
\usage{
extract_running_sim(casestudies)
}
\arguments{
\item{casestudies}{Shiny input casestudies selector. See \link{casestudiesSelect}.}
}
\description{
Simulations are currently php1, hypoD3 and hypopara. Takes
casestudies as input given by the \link{casestudiesSelect} module.
}
