% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{generate_edges}
\alias{generate_edges}
\title{CardioRenal Edges Generator}
\usage{
generate_edges(
  components,
  organs,
  regulations,
  diseases,
  organs_edges_size,
  hormones_edges_size
)
}
\arguments{
\item{components}{Shiny input CardioRenal component selector. See \link{networkOptions}.}

\item{organs}{Shiny input to toggle organs display. See \link{networkOptions}.}

\item{regulations}{Shiny input to toggle hormone display. See \link{networkOptions}.}

\item{diseases}{Shiny input disease selector. See \link{diseaseSelect}.}

\item{organs_edges_size}{Shiny input for organs edges size. See \link{networkOptions}.}

\item{hormones_edges_size}{Shiny input for hormones edges size. See \link{networkOptions}.}
}
\description{
Generate edges for the CardioRenal network
}
