% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{generate_userFields}
\alias{generate_userFields}
\title{Generate user fields}
\usage{
generate_userFields(casestudies, sliderCasestudies)
}
\arguments{
\item{casestudies}{Shiny input casestudy selector. See \link{casestudiesSelect}.}

\item{sliderCasestudies}{Shiny slider input related to the current casestudy severity.
See \link{plotBox}.}
}
\description{
Use inside in the \link{userInfo}. Function that helps in generating 4 users fields,
image, stat1, stat2 and stat3, so as to reinject them in the header userMenu
}
