% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{make_plot_hypoD3}
\alias{make_plot_hypoD3}
\title{Produce plots related to vitamin D3 deficiency (hypoD3)}
\usage{
make_plot_hypoD3(sliderVal, isMobile)
}
\arguments{
\item{sliderVal}{Shiny slider input related to the current disease severity.
See \link{plotBox}.}

\item{isMobile}{Shiny input useful to scale elements based on the device screen size.}
}
\description{
Use inside the \link{plotBox} module.
}
