% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkModule.R
\name{networkCaPO4}
\alias{networkCaPO4}
\title{CaPO4 Network server module}
\usage{
networkCaPO4(input, output, session)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny Outputs}

\item{session}{Session object.}
}
\description{
Create a CaPO4 network
}
