% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxModule.R
\name{plotBox}
\alias{plotBox}
\title{plot box server module}
\usage{
plotBox(input, output, session, diseases, help)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny Outputs}

\item{session}{Session object.}

\item{diseases}{Shiny input disease selector. See \link{diseaseSelect}.}

\item{help}{Help input.}
}
\description{
Create modals, alerts, ...
}
