% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxModule.R
\name{plotBox}
\alias{plotBox}
\title{plot box server module}
\usage{
plotBox(input, output, session)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny Outputs}

\item{session}{Session object.}
}
\description{
Create modals, alerts, ...
}
