% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_CaPO4Sim.R
\name{run_CaPO4Sim}
\alias{run_CaPO4Sim}
\title{Launch the virtual patient simulator}
\usage{
run_CaPO4Sim(context = c("introduction", "virtual-patient"))
}
\arguments{
\item{context}{Choose between \code{c("introduction", "virtual-patient")}.}
}
\description{
Unleash the virtual patient simulator
}
\examples{

if (interactive()) {
 run_CaPO4Sim(context = "introduction")
 run_CaPO4Sim(context = "virtual-patient")
}
}
