% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userInfoModule.R
\name{userInfo}
\alias{userInfo}
\title{CaPO4 user info server module}
\usage{
userInfo(input, output, session, casestudies, sliderCasestudies, help)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny Outputs}

\item{session}{Session object.}

\item{casestudies}{Shiny input casestudies selector. See \link{casestudiesSelect}.}

\item{sliderCasestudies}{Shiny input casestudy severity selector. See \link{plotBox}.}

\item{help}{Help input.}
}
\description{
Create a CaPO4 user info card
}
