% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userInfoModule.R
\name{userInfo}
\alias{userInfo}
\title{CaPO4 user info server module}
\usage{
userInfo(input, output, session)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny Outputs}

\item{session}{Session object.}
}
\description{
Create a CaPO4 user info card
}
